+++ source/extensions/omni.isaac.orbit/docs/CHANGELOG.rst
@@ -659,6 +659,14 @@ Fixed
 Added
 ^^^^^
 
+<<<<<<< HEAD
+=======
+<<<<<<< HEAD
+* Added new IMU sensor implementation that directly accessess the physx view :class:`omni.isaac.orbit.sensors.IMU`. The
+  sensor comes with a configuration class :class:`omni.isaac.orbit.sensors.IMUCfg` and data class
+  :class:`omni.isaac.orbit.sensors.IMUData`.
+=======
+>>>>>>> b88bf40944ca6243bc9a56ed4a9fb0ffad8ed655
 * Added support for using articulations as visualization markers. This disables all physics APIs from
   the articulation and allows the user to use it as a visualization marker. It is useful for creating
   visualization markers for the end-effectors or base of the robot.
@@ -670,6 +678,10 @@ Fixed
   :class:`omni.isaac.orbit.markers.VisualizationMarkers` class. Earlier, the properties were applied on
   the XForm prim instead of the Mesh prim.
 
+<<<<<<< HEAD
+=======
+>>>>>>> main
+>>>>>>> b88bf40944ca6243bc9a56ed4a9fb0ffad8ed655
 
 0.10.5 (2023-12-18)
 ~~~~~~~~~~~~~~~~~~~